third_party/WebKit/LayoutTests/fast/css/pseudo-valid-unapplied-expected.txt

This test performs a check for the :valid CSS selector on various input elements and other elements where it shouldn't be applied.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getBackgroundColor('input-text-readonly') is normalColor
PASS getBackgroundColor('input-text-disabled') is normalColor
PASS getBackgroundColor('input-button') is normalColor
PASS getBackgroundColor('input-reset') is normalColor
PASS getBackgroundColor('input-hidden') is normalColor
PASS getBackgroundColor('input-image') is normalColor
PASS getBackgroundColor('object') is normalColor
PASS getBackgroundColor('button-button') is normalColor
PASS getBackgroundColor('button-reset') is normalColor
PASS getBackgroundColor('progress') is normalColor
PASS getBackgroundColor('meter') is normalColor
PASS getBackgroundColor('option') is normalColor
PASS successfullyParsed is true

TEST COMPLETE
